requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
flask>=3.0.0
flask-session>=0.5.0
selenium>=4.15.0
webdriver-manager>=4.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
